feat: restrict appointmentbooking to logged-in users
